feat(website)!: fold the playground page into a `main-editor` story
