+++ website/package.json
@@ -23,8 +23,6 @@
     "astro-theme-toggle": "^0.8.1",
     "astrobook": "^0.13.3",
     "clsx": "^2.1.1",
-    "lz-string": "^1.5.0",
-    "nuqs": "^2.9.5",
     "react": "^19.2.8",
     "react-dom": "^19.2.8"
   },
