+++ website/src/pages/index.astro
@@ -24,12 +24,6 @@ const iconButtonClass =
         </div>
         <div class="flex items-center gap-2">
           <a
-            href="/playground"
-            class="text-sm font-medium text-stone-500 underline-offset-4 transition-colors hover:text-stone-900 hover:underline dark:text-stone-400 dark:hover:text-stone-100"
-          >
-            Playground
-          </a>
-          <a
             href="/stories"
             class="mr-1 text-sm font-medium text-stone-500 underline-offset-4 transition-colors hover:text-stone-900 hover:underline dark:text-stone-400 dark:hover:text-stone-100"
           >
