--- website/src/pages/playground.astro
@@ -1,41 +1,0 @@
----
-import { Toggle } from 'astro-theme-toggle/components'
-
-import Base from '../layouts/base.astro'
-import Playground from '../playground/playground.tsx'
-
-const iconButtonClass =
-  'flex h-8 w-8 cursor-pointer items-center justify-center rounded-full border border-stone-200/80 bg-white/70 text-stone-500 shadow-sm backdrop-blur transition-colors hover:bg-white hover:text-stone-900 dark:border-stone-700/70 dark:bg-stone-900/70 dark:text-stone-400 dark:hover:bg-stone-800 dark:hover:text-stone-100'
----
-
-<Base title="Meowdown Playground">
-  <main class="flex h-dvh flex-col text-stone-600 dark:bg-stone-950">
-    <nav class="flex shrink-0 items-center gap-3 px-4 py-3 sm:px-6">
-      <a href="/" class="flex items-center gap-2">
-        <span
-          class="flex h-7 w-7 items-center justify-center rounded-lg bg-yellow-100 text-sm shadow-sm shadow-amber-500/30"
-        >
-          🐱
-        </span>
-        <span class="font-semibold tracking-tight text-stone-900 dark:text-white">Meowdown</span>
-      </a>
-      <span class="text-sm text-stone-400 dark:text-stone-600">/</span>
-      <span class="text-sm font-medium text-stone-900 dark:text-white">Playground</span>
-      <div class="ml-auto flex items-center gap-2">
-        <a
-          href="/stories"
-          class="mr-1 text-sm font-medium text-stone-500 underline-offset-4 transition-colors hover:text-stone-900 hover:underline dark:text-stone-400 dark:hover:text-stone-100"
-        >
-          Stories
-        </a>
-        <Toggle
-          class={iconButtonClass}
-          style="padding: 8px;"
-          aria-label="Toggle color theme"
-          title="Toggle color theme"
-        />
-      </div>
-    </nav>
-    <Playground client:only="react" />
-  </main>
-</Base>
