+++ website/src/stories/main-editor.stories.ts
@@ -1,0 +1,7 @@
+import { MainEditor as MainEditorComponent } from './main-editor.tsx'
+
+export default {
+  component: MainEditorComponent,
+}
+
+export const MainEditor = {}
